refactor: 카테고리 서비스 -> 카테고리 조회 쿼리로 변경
