+++ src/main/java/com/meoguri/linkocean/domain/category/service/FavoriteCategoryServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.meoguri.linkocean.domain.category.entity.FavoriteCategory;
 import com.meoguri.linkocean.domain.category.persistence.FavoriteCategoryRepository;
+import com.meoguri.linkocean.domain.category.persistence.FindCategoriesByNamesQuery;
 import com.meoguri.linkocean.domain.category.service.dto.AddFavoriteCategoriesCommand;
 
 import lombok.RequiredArgsConstructor;
@@ -16,8 +17,8 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class FavoriteCategoryServiceImpl implements FavoriteCategoryService {
 
+	private final FindCategoriesByNamesQuery findCategoriesByNamesQuery;
 	private final FavoriteCategoryRepository favoriteCategoryRepository;
-	private final CategoryService categoryService;
 
 	@Override
 	public void addFavoriteCategories(final AddFavoriteCategoriesCommand command) {
@@ -25,7 +26,7 @@ public class FavoriteCategoryServiceImpl implements FavoriteCategoryService {
 		final List<String> categoryNames = command.getCategoryNames();
 
 		//TODO - Profile 서비스 구현 후 profileId 로 Profile 조회 하여 채우기 지금은 임시로 null
-		categoryService.findByNames(categoryNames)
+		findCategoriesByNamesQuery.findByNames(categoryNames)
 			.forEach(c -> favoriteCategoryRepository.save(new FavoriteCategory(null, c)));
 	}
 }
