+++ src/main/java/com/meoguri/linkocean/domain/profile/persistence/GetProfileByUserIdQuery.java
@@ -1,7 +1,5 @@
 package com.meoguri.linkocean.domain.profile.persistence;
 
-import org.springframework.transaction.annotation.Transactional;
-
 import com.meoguri.linkocean.annotation.Query;
 import com.meoguri.linkocean.domain.profile.entity.Profile;
 import com.meoguri.linkocean.exception.LinkoceanRuntimeException;
@@ -9,7 +7,6 @@ import com.meoguri.linkocean.exception.LinkoceanRuntimeException;
 import lombok.RequiredArgsConstructor;
 
 @Query
-@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class GetProfileByUserIdQuery {
