+++ src/test/java/com/meoguri/linkocean/domain/category/persistence/FindCategoriesByNamesQueryTest.java
@@ -1,4 +1,6 @@
-package com.meoguri.linkocean.domain.category.service;
+package com.meoguri.linkocean.domain.category.persistence;
+
+import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
 
@@ -12,8 +14,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.meoguri.linkocean.domain.category.persistence.CategoryRepository;
-import com.meoguri.linkocean.domain.category.service.dto.AddFavoriteCategoriesCommand;
+import com.meoguri.linkocean.domain.category.entity.Category;
 
 @Disabled
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
@@ -21,10 +22,10 @@ import com.meoguri.linkocean.domain.category.service.dto.AddFavoriteCategoriesCo
 @Sql("classpath:db/sql/InsertCategories.sql")
 @Transactional
 @SpringBootTest
-class FavoriteCategoryServiceImplTest {
+class FindCategoriesByNamesQueryTest {
 
 	@Autowired
-	private FavoriteCategoryService favoriteCategoryService;
+	private FindCategoriesByNamesQuery query;
 
 	@Autowired
 	private CategoryRepository categoryRepository;
@@ -32,19 +33,18 @@ class FavoriteCategoryServiceImplTest {
 	@AfterAll
 	void setUp() {
 		categoryRepository.deleteAllInBatch();
-
-		//TODO - User, Profile 하나씩 추가
 	}
 
-	// TODO - 프로필 서비스 구현 이후 완성
 	@Test
-	void 선호_카테고리_추가_성공() {
+	void 이름_목록으로_조회_성공() {
 		//given
-		favoriteCategoryService.addFavoriteCategories(
-			new AddFavoriteCategoriesCommand(
-				1L,
-				List.of("자기계발", "인문")
-			)
-		);
+		final String name1 = "자기계발";
+		final String name2 = "인문";
+
+		//when
+		final List<Category> categories = query.findByNames(List.of(name1, name2));
+
+		//then
+		assertThat(categories).extracting(Category::getName).containsExactlyInAnyOrder(name1, name2);
 	}
 }
