+++ src/test/java/com/meoguri/linkocean/domain/category/service/FavoriteCategoryServiceTest.java
@@ -1,7 +1,5 @@
 package com.meoguri.linkocean.domain.category.service;
 
-import static org.assertj.core.api.Assertions.*;
-
 import java.util.List;
 
 import org.junit.jupiter.api.AfterAll;
@@ -14,8 +12,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.meoguri.linkocean.domain.category.entity.Category;
 import com.meoguri.linkocean.domain.category.persistence.CategoryRepository;
+import com.meoguri.linkocean.domain.category.service.dto.AddFavoriteCategoriesCommand;
 
 @Disabled
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
@@ -23,10 +21,10 @@ import com.meoguri.linkocean.domain.category.persistence.CategoryRepository;
 @Sql("classpath:db/sql/InsertCategories.sql")
 @Transactional
 @SpringBootTest
-class CategoryServiceImplTest {
+class FavoriteCategoryServiceTest {
 
 	@Autowired
-	private CategoryService categoryService;
+	private FavoriteCategoryService favoriteCategoryService;
 
 	@Autowired
 	private CategoryRepository categoryRepository;
@@ -34,18 +32,19 @@ class CategoryServiceImplTest {
 	@AfterAll
 	void setUp() {
 		categoryRepository.deleteAllInBatch();
+
+		//TODO - User, Profile 하나씩 추가
 	}
 
+	// TODO - 프로필 서비스 구현 이후 완성
 	@Test
-	void 이름_목록으로_조회_성공() {
+	void 선호_카테고리_추가_성공() {
 		//given
-		final String name1 = "자기계발";
-		final String name2 = "인문";
-
-		//when
-		final List<Category> categories = categoryService.findByNames(List.of(name1, name2));
-
-		//then
-		assertThat(categories).extracting(Category::getName).containsExactlyInAnyOrder(name1, name2);
+		favoriteCategoryService.addFavoriteCategories(
+			new AddFavoriteCategoriesCommand(
+				1L,
+				List.of("자기계발", "인문")
+			)
+		);
 	}
 }
